IA/requirements2.txt
openai
pillow
flask
torch